add --build-optimizer
https://blog.angular.io/the-past-present-and-future-of-the-angular-cli-13cf55e455f8

With each release, we work hard to try to improve the quality of the code that the CLI generates with production builds. With this release we’re leveraging the latest capabilities from Webpack 3, including scope hoisting.
We’ve also included a beta of a new tool we are working on to help with removing code that you aren’t reliant on called our build optimizer. You can give this flag a try with:
ng build --prod --build-optimizer
From our own work on angular.io, we saw a 52% size reduction in our production build.